minor fixes

+++ genial/src/main/java/com/genial/demo/shared/ProductCreate.java
@@ -1,10 +1,12 @@
 package com.genial.demo.shared;
 
+import java.math.BigDecimal;
+
 public record ProductCreate(
                 String name,
                 String description,
                 String sector,
-                Double value,
+                BigDecimal value,
                 Integer quantidade) {
 
 }
